% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookiemonster-package.R
\docType{package}
\name{cookiemonster-package}
\alias{cookiemonster-package}
\title{A package to create cookiemonster lineup plot using nullabor}
\description{
A package to create cookiemonster lineup plot using nullabor
}
