% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{delete_cookies}
\alias{delete_cookies}
\title{Delete Cookies}
\usage{
delete_cookies(domain, jar = default_jar(), ask = TRUE)
}
\arguments{
\item{domain}{The domain for which the cookies should be deleted.}

\item{jar}{A character string of the path to the cookie jar (the default is
to use \code{default_jar()} to get a suitable directory).}

\item{ask}{A logical value indicating whether the user should be asked to confirm the deletion.}
}
\value{
Nothing
}
\description{
Delete Cookies
}
\examples{
delete_cookies("example.com", ask = FALSE)
}
