% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{get_cookies}
\alias{get_cookies}
\title{Retrieve cookies from a jar}
\usage{
get_cookies(
  domain,
  jar = default_jar(),
  as = c("data.frame", "string", "vector")
)
}
\arguments{
\item{domain}{A character string of the domain to retrieve cookies for.}

\item{jar}{A character string of the path to the cookie jar (the default is
to use \code{default_jar()} to get a suitable directory).}

\item{as}{A character string of the type of output to return.}
}
\value{
Depending on the value of \code{as}, returns either a data frame, a
character string, or a named vector.
}
\description{
Imagine you're reaching into a magical jar overflowing with those scrumptious
digital delights from websites you've visited. The flavour? Up to you! Just
select your desired output format.
}
\details{
The function returns cookies in one of three formats:

\itemize{
\item{\strong{data.frame:}} is how cookies are stored internally and can be used for manual inspection.
\item{\strong{string:}} is used by \code{curl} and \code{httr2}.
\item{\strong{vector:}} is used by \code{httr}.
}

See \code{vignette("cookies", "cookiemonster")} for how to use cookies with
these packages.
}
\note{
Your cookies are saved in an encrypted file. See \link{encrypt_vec} for
more info.
}
\examples{
\dontrun{get_cookies("https://hb.cran.dev")}  # Reach into your cookie jar and enjoy!
}
\seealso{
\code{\link{add_cookies}}
}
