% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.r
\name{get_cookies_from_browser}
\alias{get_cookies_from_browser}
\title{Get Cookies from Browser}
\usage{
get_cookies_from_browser(browser = "Firefox")
}
\arguments{
\item{browser}{A character string specifying the browser to get cookies from.
Defaults to "Firefox".}
}
\value{
A tibble containing the cookies from the specified browser.
}
\description{
Retrieves all cookies from a Browser (currently works with Firefox only).
}
\examples{
\dontrun{
get_cookies_from_browser("Firefox")
}
}
