% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monster_raster}
\alias{monster_raster}
\title{Raster values of cookie monster image.}
\format{An object of class \code{array} of dimension 240 x 860 x 4.}
\usage{
monster_raster
}
\description{
A three-dimensional array used to plot the cookie monster image.
}
\keyword{datasets}
