% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.r
\name{add_cookies}
\alias{add_cookies}
\title{Add Cookies to the Browser}
\usage{
add_cookies(cookiefile, cookiestring, domain = NULL)
}
\arguments{
\item{cookiefile}{A character string indicating the path to the cookie file.}

\item{cookiestring}{A character string representing the cookie in string
format.}

\item{domain}{An optional parameter that specifies the host/domain. It's only
used when \code{cookiestring} is provided.}
}
\value{
No explicit return. Instead, this function stores the cookies using
the \code{store_cookies} function.
}
\description{
This function allows you to add browser cookies to the cookie storage. It can
work with either a cookie file or a direct cookie string (e.g., copied from a
CURL call). But remember, just like in real life, you can't have your cookie
and eat it too - pick only one!
}
\note{
You can't provide both a cookiefile and a cookiestring at the same
time. That's like trying to dunk two cookies in a tiny cup of milk!

Your cookies are saved in an encrypted file. See \link{encrypt_vec} for
more info.
}
\examples{
# Using a cookie file:
# Using the example `cookies.txt` file which contains the cookie details.
add_cookies(cookiefile = system.file("extdata", "cookies.txt", package = "cookiemonster"))

# Using a cookie string:
add_cookies(cookiestring = "username=johndoe; password=secret", domain = "www.example.com")

}
\seealso{
\code{\link{store_cookies}}
}
