% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.r
\name{store_cookies}
\alias{store_cookies}
\title{Store cookies in a jar}
\usage{
store_cookies(cookies, jar = default_jar())
}
\arguments{
\item{cookies}{A data frame of cookies}

\item{jar}{The directory to store the cookies in. Defaults to \code{default_jar()}.}
}
\description{
Store cookies in a jar
}
\examples{
store_cookies(cookies)
}
